move supported chains
