+++ crates/cli/commands/src/db/mod.rs
@@ -161,7 +161,7 @@ impl<C: ChainSpecParser<ChainSpec: EthChainSpec + EthereumHardforks>> Command<C>
 mod tests {
     use super::*;
     use reth_ethereum_cli::chainspec::EthChainSpecParser;
-    use reth_node_core::args::utils::{SUPPORTED_CHAINS};
+    use reth_ethereum_cli::chainspec::SUPPORTED_CHAINS;
     use std::path::Path;
 
     #[test]
