+++ crates/cli/commands/src/import.rs
@@ -232,7 +232,7 @@ where
 mod tests {
     use super::*;
     use reth_ethereum_cli::chainspec::EthChainSpecParser;
-    use reth_node_core::args::utils::{SUPPORTED_CHAINS};
+    use reth_ethereum_cli::chainspec::SUPPORTED_CHAINS;
 
     #[test]
     fn parse_common_import_command_chain_args() {
